requirements.txt
aws-cdk-lib==2.46.0
constructs>=10.0.0,<11.0.0
boto3
botocore
simplejson
aws-cdk.aws-lambda-python-alpha

